+++ tests/acceptance/pods-destroy-test.js
@@ -91,7 +91,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['controller', 'foo', '--pod'];
     var files       = [
       'app/pods/foo/controller.js',
-      'tests/unit/controllers/foo-test.js'
+      'tests/unit/pods/foo/controller-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -101,7 +101,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['controller', 'foo/bar', '--pod'];
     var files       = [
       'app/pods/foo/bar/controller.js',
-      'tests/unit/controllers/foo/bar-test.js'
+      'tests/unit/pods/foo/bar/controller-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -112,7 +112,7 @@ describe('Acceptance: ember destroy pod', function() {
     var files       = [
       'app/pods/components/x-foo/component.js',
       'app/pods/components/x-foo/template.hbs',
-      'tests/unit/components/x-foo-test.js'
+      'tests/unit/pods/components/x-foo/component-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -142,7 +142,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['model', 'foo', '--pod'];
     var files       = [
       'app/pods/foo/model.js',
-      'tests/unit/models/foo-test.js'
+      'tests/unit/pods/foo/model-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -152,7 +152,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['model', 'foo/bar', '--pod'];
     var files       = [
       'app/pods/foo/bar/model.js',
-      'tests/unit/models/foo/bar-test.js'
+      'tests/unit/pods/foo/bar/model-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -163,7 +163,7 @@ describe('Acceptance: ember destroy pod', function() {
     var files       = [
       'app/pods/foo/route.js',
       'app/pods/foo/template.hbs',
-      'tests/unit/routes/foo-test.js'
+      'tests/unit/pods/foo/route-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -174,7 +174,7 @@ describe('Acceptance: ember destroy pod', function() {
     var files       = [
       'app/pods/foo/route.js',
       'app/pods/foo/template.hbs',
-      'tests/unit/routes/foo-test.js'
+      'tests/unit/pods/foo/route-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files)
@@ -190,7 +190,7 @@ describe('Acceptance: ember destroy pod', function() {
     var files       = [
       'app/pods/foos/route.js',
       'app/pods/foos/template.hbs',
-      'tests/unit/routes/foos-test.js'
+      'tests/unit/pods/foos/route-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files)
@@ -206,7 +206,7 @@ describe('Acceptance: ember destroy pod', function() {
     var files       = [
       'app/pods/index/route.js',
       'app/pods/index/template.hbs',
-      'tests/unit/routes/index-test.js'
+      'tests/unit/pods/index/route-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -217,7 +217,7 @@ describe('Acceptance: ember destroy pod', function() {
     var files       = [
       'app/pods/basic/route.js',
       'app/pods/basic/template.hbs',
-      'tests/unit/routes/basic-test.js'
+      'tests/unit/pods/basic/route-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -227,9 +227,9 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['resource', 'foo', '--pod'];
     var files       = [
       'app/pods/foo/model.js',
-      'tests/unit/models/foo-test.js',
+      'tests/unit/pods/foo/model-test.js',
       'app/pods/foo/route.js',
-      'tests/unit/routes/foo-test.js',
+      'tests/unit/pods/foo/route-test.js',
       'app/pods/foo/template.hbs'
     ];
 
@@ -245,9 +245,9 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['resource', 'foos', '--pod'];
     var files       = [
       'app/pods/foo/model.js',
-      'tests/unit/models/foo-test.js',
+      'tests/unit/pods/foo/model-test.js',
       'app/pods/foos/route.js',
-      'tests/unit/routes/foos-test.js',
+      'tests/unit/pods/foos/route-test.js',
       'app/pods/foos/template.hbs'
     ];
 
@@ -277,7 +277,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['view', 'foo', '--pod'];
     var files       = [
       'app/pods/foo/view.js',
-      'tests/unit/views/foo-test.js'
+      'tests/unit/pods/foo/view-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -287,7 +287,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['view', 'foo/bar', '--pod'];
     var files       = [
       'app/pods/foo/bar/view.js',
-      'tests/unit/views/foo/bar-test.js'
+      'tests/unit/pods/foo/bar/view-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -345,7 +345,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['serializer', 'foo', '--pod'];
     var files       = [
       'app/pods/foo/serializer.js',
-      'tests/unit/serializers/foo-test.js'
+      'tests/unit/pods/foo/serializer-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -355,7 +355,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['serializer', 'foo/bar', '--pod'];
     var files       = [
       'app/pods/foo/bar/serializer.js',
-      'tests/unit/serializers/foo/bar-test.js'
+      'tests/unit/pods/foo/bar/serializer-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -365,7 +365,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['transform', 'foo', '--pod'];
     var files       = [
       'app/pods/foo/transform.js',
-      'tests/unit/transforms/foo-test.js'
+      'tests/unit/pods/foo/transform-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);
@@ -375,7 +375,7 @@ describe('Acceptance: ember destroy pod', function() {
     var commandArgs = ['transform', 'foo/bar', '--pod'];
     var files       = [
       'app/pods/foo/bar/transform.js',
-      'tests/unit/transforms/foo/bar-test.js'
+      'tests/unit/pods/foo/bar/transform-test.js'
     ];
 
     return assertDestroyAfterGenerate(commandArgs, files);